Add exports and Gh Actions, removed packages
