+++ .github/workflows/test.yaml
@@ -1,0 +1,33 @@
+name: Build test
+on:
+  push:
+    branches:
+      - master
+
+jobs:
+  nodejs:
+    strategy:
+      matrix:
+        nodejs:
+          - lts/*
+          - latest
+        os:
+          - ubuntu
+          - windows
+          - macos
+    runs-on: ${{ matrix.os }}-latest
+    steps:
+      - uses: actions/checkout@v3
+        name: Get source
+        with:
+          submodules: true
+
+      - uses: actions/setup-node@v3
+        name: Setup Node.js ${{ matrix.nodejs }}
+        with:
+          node-version: ${{ matrix.nodejs }}
+
+      - name: Install depencies
+        run: npm install
+
+      - run: npm run rebuild
