--- .travis.yml
@@ -1,11 +1,0 @@
-language: node_js
-node_js:
-  - "6"
-  - "8"
-  - "10"
-addons:
-  apt:
-    sources:
-      - ubuntu-toolchain-r-test
-    packages:
-      - g++
